requirements.txt
certifi==2019.6.16
joblib==0.13.2
mdlp-discretization==0.3.2
mkl-fft==1.0.12
mkl-random==1.0.2
numpy==1.16.4
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
